requirements.txt
asgiref==3.6.0
astroid==2.15.3
certifi==2022.12.7
charset-normalizer==3.1.0
dill==0.3.6
Django==4.2
djangorestframework==3.14.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
numpy==1.24.2
pandas==2.0.0
platformdirs==3.2.0
pylint==2.17.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
schedule==1.2.0
six==1.16.0
sqlparse==0.4.3
tomlkit==0.11.7
typing==3.7.4.3
tzdata==2023.3
urllib3==1.26.15
vvspy==1.2.0
wrapt==1.15.0
